move question names to types file
