+++ src/Types/Questions.ts
@@ -1,0 +1,11 @@
+export type QuestionNames =
+  | 'zipcode'
+  | 'healthInsurance'
+  | 'householdSize'
+  | 'householdData'
+  | 'hasExpenses'
+  | 'householdAssets'
+  | 'hasBenefits'
+  | 'acuteHHConditions'
+  | 'referralSource'
+  | 'signUpInfo';
